Add support for username file
This commit expands the variations of how we find users to use for the montage
to include a list of usernames. Because we are not retrieving these usernames
from the API ourself, the script will fail as soon as any invalid username is
encountered. Additionally, it will determine the montage image name based on
the input file provided, which can be overridden.